requirements.txt
aiogram~=3.10.0
python-dotenv~=1.0.1
loguru~=0.7.2
sqlmodel~=0.0.21
aiohttp[speedups]~=3.9.0
ngrok~=1.4.0
